+++ default.nix
@@ -1,0 +1,3 @@
+{ pkgs ? import <nixpkgs> {} }:
+
+pkgs.haskellPackages.callCabal2nix "exp-dvector-prediction-wains" ./. {}