+++ src/ALife/Creatur/Wain/DVector/Prediction/ResponseTweaker.hs
@@ -11,8 +11,8 @@
 --
 ------------------------------------------------------------------------
 {-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE DeriveGeneric  #-}
+{-# LANGUAGE TypeFamilies   #-}
 module ALife.Creatur.Wain.DVector.Prediction.ResponseTweaker
   (
     ResponseTweaker(..),
@@ -22,20 +22,31 @@ module ALife.Creatur.Wain.DVector.Prediction.ResponseTweaker
     -- makeResponseSimilar
   ) where
 
-import qualified ALife.Creatur.Genetics.BRGCWord8 as W8
-import ALife.Creatur.Genetics.Diploid (Diploid)
-import ALife.Creatur.Wain.GeneticSOM (Difference, Tweaker(..))
-import ALife.Creatur.Wain.PlusMinusOne (adjustPM1Vector)
-import  ALife.Creatur.Wain.Pretty (Pretty)
-import ALife.Creatur.Wain.Response (Response(..), labelSimilarity)
-import ALife.Creatur.Wain.Statistics (Statistical(..), prefix)
-import ALife.Creatur.Wain.DVector.Prediction.Action (Action,
-  actionDiff, makeActionSimilar)
-import ALife.Creatur.Wain.UnitInterval (UIDouble)
-import ALife.Creatur.Wain.Weights (Weights, weightedSum)
-import Control.DeepSeq (NFData)
-import Data.Serialize (Serialize)
-import GHC.Generics (Generic)
+import qualified ALife.Creatur.Genetics.BRGCWord8             as W8
+import           ALife.Creatur.Genetics.Diploid
+    (Diploid)
+import           ALife.Creatur.Wain.DVector.Prediction.Action
+    (Action, actionDiff, makeActionSimilar)
+import           ALife.Creatur.Wain.GeneticSOM
+    (Difference, Tweaker (..))
+import           ALife.Creatur.Wain.PlusMinusOne
+    (adjustPM1Vector)
+import           ALife.Creatur.Wain.Pretty
+    (Pretty)
+import           ALife.Creatur.Wain.Response
+    (Response (..), labelSimilarity)
+import           ALife.Creatur.Wain.Statistics
+    (Statistical (..), prefix)
+import           ALife.Creatur.Wain.UnitInterval
+    (UIDouble)
+import           ALife.Creatur.Wain.Weights
+    (Weights, weightedSum)
+import           Control.DeepSeq
+    (NFData)
+import           Data.Serialize
+    (Serialize)
+import           GHC.Generics
+    (Generic)
 
 -- | @'ResponseTweaker'@ constructs an object which is
 --   responsible for comparing and adjusting response patterns.
@@ -44,8 +55,8 @@ import GHC.Generics (Generic)
 --   reference the action type @a@. As a result, @ResponseTweaker@ has
 --   to have a type parameter @a@, even though it is not used.
 data ResponseTweaker = ResponseTweaker Weights
-  deriving (Eq, Show, Pretty, Generic, NFData, Serialize, W8.Genetic,
-            Diploid)
+  deriving (Eq, Show, Read, Pretty, Generic, NFData, Serialize,
+            W8.Genetic, Diploid)
 
 instance Tweaker ResponseTweaker where
   type Pattern ResponseTweaker = Response Action