+++ test/ALife/Creatur/Wain/DVector/Prediction/ActionQC.hs
@@ -16,13 +16,23 @@ module ALife.Creatur.Wain.DVector.Prediction.ActionQC
     test
   ) where
 
-import ALife.Creatur.Wain.TestUtils (prop_serialize_round_trippable,
-  prop_genetic_round_trippable, prop_diploid_identity)
-import ALife.Creatur.Wain.DVector.Prediction.Action
-import ALife.Creatur.Wain.DVector.Double (minDouble, maxDouble)
-import Test.Framework (Test, testGroup)
-import Test.Framework.Providers.QuickCheck2 (testProperty)
-import Test.QuickCheck
+import           ALife.Creatur.Wain.DVector.Double
+    (maxDouble, minDouble)
+import           ALife.Creatur.Wain.DVector.Prediction.Action
+import           ALife.Creatur.Wain.TestUtils
+    ( prop_diploid_expressable
+    , prop_diploid_identity
+    , prop_diploid_readable
+    , prop_genetic_round_trippable
+    , prop_makeSimilar_works
+    , prop_serialize_round_trippable
+    , prop_show_read_round_trippable
+    )
+import           Test.Framework
+    (Test, testGroup)
+import           Test.Framework.Providers.QuickCheck2
+    (testProperty)
+import           Test.QuickCheck
 
 instance Arbitrary Action where
   arbitrary = fmap mkAction arbitrary
@@ -53,8 +63,19 @@ test = testGroup "ALife.Creatur.Wain.DVector.Prediction.ActionQC"
       (prop_serialize_round_trippable :: Action -> Property),
     testProperty "prop_genetic_round_trippable - Action"
       (prop_genetic_round_trippable (==) :: Action -> Property),
+    -- testProperty "prop_genetic_round_trippable2 - Action"
+    --   (prop_genetic_round_trippable2
+    --    :: Int -> [Word8] -> Action -> Property),
     testProperty "prop_diploid_identity - Action"
       (prop_diploid_identity (==) :: Action -> Property),
+    testProperty "prop_show_read_round_trippable - Action"
+      (prop_show_read_round_trippable (==) :: Action -> Property),
+    testProperty "prop_diploid_expressable - Action"
+      (prop_diploid_expressable :: Action -> Action -> Property),
+    testProperty "prop_diploid_readable - Action"
+      (prop_diploid_readable :: Action -> Action -> Property),
+    testProperty "prop_makeSimilar_works"
+      (prop_makeSimilar_works actionDiff makeActionSimilar),
     testProperty "prop_action_diff_is_reflexive"
       prop_action_diff_is_reflexive,
     testProperty "prop_action_diff_can_be_zero"