+++ ChangeLog.md
@@ -1,7 +1,14 @@
 # Changelog for exp-uivector-prediction-wains
 
-1.0.9 Upgraded to new version of creatur-wains.
-1.0.8 Upgraded to lts-12.16.
-1.0.7 Revamped to work with Stack v1.7.1.
+1.0.11 Revamped to use creatur-dvector-wains 1.1.3 (which uses SGM4).
+
+1.0.10 Fixed minor bug in extendBoundaries.
+       Added more tests.
+
+1.0.9  Upgraded to new version of creatur-wains.
+
+1.0.8  Upgraded to lts-12.16.
+
+1.0.7  Revamped to work with Stack v1.7.1.
 
 ## Unreleased changes